re organise file structure to remove unnecessary nesting and update import filepaths
