+++ src/components/CommandAdd.js
@@ -3,7 +3,7 @@ import Row from 'react-bootstrap/Row'
 import Col from 'react-bootstrap/Col'
 import Button from 'react-bootstrap/Button'
 import { Typeahead } from 'react-bootstrap-typeahead';
-import { availableCommands } from '../../utils/constant';
+import { availableCommands } from '../utils/constant';
 
 /** CommandAdd.
  *
