+++ src/components/RenderedCube.js
