+++ src/pages/MainPage.js
@@ -1,10 +1,10 @@
 import Container from 'react-bootstrap/Container';
-import NavigationBar from '../../components/NavigationBar/NavigationBar.js';
-import ScenePane from '../../panes/ScenePane/ScenePane'
-import CommandPane from '../../panes/CommandPane/CommandPane.js'
+import NavigationBar from '../components/NavigationBar';
+import ScenePane from '../pages/ScenePane.js'
+import CommandPane from './CommandPane.js'
 import Col from 'react-bootstrap/Col'
 import Row from 'react-bootstrap/Row'
-import cubeMachine from '../../machines/cubeMachine.js'
+import cubeMachine from '../machines/cubeMachine.js'
 import { useMachine } from '@xstate/react'
